+++ examples/ASP.Net/ASP.Net Web Application/HangfireManagement/Expedited.cs
@@ -1,10 +1,9 @@
-﻿using Hangfire;
+﻿using System;
+using System.ComponentModel;
+using Hangfire;
 using Hangfire.Dashboard.Management.v2.Metadata;
 using Hangfire.Dashboard.Management.v2.Support;
 using Hangfire.Server;
-
-using System;
-using System.ComponentModel;
 
 namespace ASP.Net_Web_Application.HangfireManagement
 {
@@ -17,40 +16,42 @@ namespace ASP.Net_Web_Application.HangfireManagement
 		[Queue("expedited")]
 		public void Job1(PerformContext context, IJobCancellationToken token,
 			[DisplayData(
-			Label = "String Input 1",
-			Description = "This is the description text for the string input with a default value and the control is disabled",
-			DefaultValue = "This is the Default Value",
-			IsDisabled = true
-		)]
-		string strInput1,
+				Label = "String Input 1",
+				Description = "This is the description text for the string input with a default value and the control is disabled",
+				DefaultValue = "This is the Default Value",
+				IsDisabled = true
+			)] string strInput1,
 
 			[DisplayData(
-			Placeholder = "This is the placeholder text",
-			Description = "This is the description text for the string input without a default value and the control is enabled"
-		)]
-		string strInput2,
+				Placeholder = "This is the placeholder text",
+				Description = "This is the description text for the string input without a default value and the control is enabled"
+			)] string strInput2,
 
 			[DisplayData(
-			Label = "DateTime Input",
-			Placeholder = "What is the date and time?",
-			DefaultValue = "1/20/2020 1:02 AM",
-			Description = "This is a date time input control"
-		)]
-		DateTime dtInput,
+				Label = "Multiline Input",
+				IsMultiLine = true,
+				Placeholder = "This is the multiline\nplaceholder text",
+				Description = "This is the description text for the multiline input without a default value where the control is enabled and not required"
+			)] string strInput3,
 
 			[DisplayData(
-			Label = "Boolean Input",
-			DefaultValue = true,
-			Description = "This is a boolean input"
-		)]
-		bool blInput,
+				Label = "DateTime Input",
+				Placeholder = "What is the date and time?",
+				DefaultValue = "1/20/2020 1:02 AM",
+				Description = "This is a date time input control"
+			)] DateTime dtInput,
 
 			[DisplayData(
-			Label = "Select Input",
-			DefaultValue = TestEnum.Test5,
-			Description = "Based on an enum object"
-		)]
-		TestEnum enumTest
+				Label = "Boolean Input",
+				DefaultValue = true,
+				Description = "This is a boolean input"
+			)] bool blInput,
+
+			[DisplayData(
+				Label = "Select Input",
+				DefaultValue = TestEnum.Test5,
+				Description = "Based on an enum object"
+			)] TestEnum enumTest
 		)
 		{
 			//Do awesome things here