+++ src/Metadata/DisplayDataAttribute.cs
@@ -51,7 +51,7 @@ namespace Hangfire.Dashboard.Management.v2.Metadata
 		public bool IsRequired { get; set; } = false;
 
 		public DisplayDataAttribute() { }
-		public DisplayDataAttribute(string label = null, string placeholder = null, string description = null, object defaultValue = null, string cssClasses = null, bool isDisabled = false, bool isRequired = false)
+		public DisplayDataAttribute(string label = null, string placeholder = null, string description = null, object defaultValue = null, string cssClasses = null, bool isDisabled = false, bool isRequired = false, bool isMultiline = false)
 		{
 			this.Label = label;
 			this.Placeholder = placeholder;
@@ -59,7 +59,8 @@ namespace Hangfire.Dashboard.Management.v2.Metadata
 			this.DefaultValue = defaultValue;
 			this.CssClasses = cssClasses;
 			this.IsDisabled = isDisabled;
-			this.IsRequired = IsRequired;
+			this.IsRequired = isRequired;
+			this.IsMultiLine = isMultiline;
 		}
 	}
 }